chapters/2025/Integration_du_developpement_ML_dans_un_pipeline_CICD_Team_D/assets/codes/requirements.txt
GitPython==3.1.44
matplotlib==3.10.0
seaborn==0.13.2
numpy==2.2.1
plotly==5.24.1
requests==2.32.3